#40 refactor asserts for reset to common class
