+++ goddo_test/test_drop_file.py
@@ -6,6 +6,7 @@ from PyQt5.QtWidgets import QLabel
 from goddo_player.utils.video_path import VideoPath
 from goddo_player.utils.url_utils import file_to_url
 from goddo_player.utils.window_util import local_to_global_pos
+from goddo_test.common_asserts import assert_state_is_blank
 from goddo_test.utils.command_widget import Command, CommandType
 from goddo_test.utils.path_util import video_folder_path, my_test_output_folder_path
 from goddo_test.utils.test_utils import wait_until, drag_and_drop
@@ -58,7 +59,7 @@ def test_drop_vid_file(app_thread, windows_container, test_file_ext):
 
     app_thread.cmd.submit_cmd(Command(CommandType.RESET))
 
-    assert_state_is_blank(app_thread)
+    assert_state_is_blank(app_thread, windows_container)
 
     app_thread.cmd.submit_cmd(Command(CommandType.LOAD_FILE, [VideoPath(file_to_url(str(save_file_path)))]))
 
@@ -95,12 +96,3 @@ def assert_after_drop(app_thread, video_path, new_total_count_expected):
     screenshot_label = item_widget.findChildren(QLabel, "screenshot")[0]
     pixmap = screenshot_label.pixmap()
     assert pixmap != videos_tab.black_pixmap
-
-
-def assert_state_is_blank(app_thread):
-    file_list_state = app_thread.mon.state.file_list
-    videos_tab = app_thread.mon.tabbed_list_window.videos_tab
-    video_tab_list_widget = videos_tab.listWidget
-    assert len(file_list_state.files) == 0
-    assert len(file_list_state.files_dict) == 0
-    assert video_tab_list_widget.count() == 0
